replace javascript-state-machine with basic singleton class
