+++ package.json
@@ -34,7 +34,6 @@
   },
   "dependencies": {
     "@tweenjs/tween.js": "^21.0.0",
-    "javascript-state-machine": "^3.1.0",
     "kontra": "^9.0.0",
     "screenfull": "^6.0.2",
     "tinymusic": "^1.0.0"
